rename vartime_pippenger to pippenger
