+++ src/backend/avx2/scalar_mul/pippenger.rs
